WIP adding base test class to be used by rest of test classes
