+++ csunplugged/tests/__init__.py
@@ -1,0 +1,4 @@
+import unittest
+
+def suite():
+	return  unittest.TestLoader().discover('topics.tests', pattern='*.py')
